PR1 prep: keep settings rename while reverting extra logic edits
